This is a major rework of Kranc-devel, which adds compatibility with Carpet fixed mesh refinement, standard Cactus boundary conditions and Lego Excision.
Changes are documented briefly in the ChangeLog file, and I will do more
documentation in the next days, along with necessary adaptions of the examples.

In order to use this, the parameter GenericFD::stencil_size has to be set
(this is necessary to treat real boundaries, since cctk_nghostzones >
stencil size is possible, e.g. for Carpet runs).

Also, it is now necessary that GF names do not coincide with their group names.
This avoids Cactus warnings messages and allows simple names for parameters associated
with standard Cactus boundary conditions.

Note: In order to allow boundary conditions which are consistent with characteristic information,
one-sided differencing will be required, which is not implemented by these changes.

So far this has only been tested by running a Kerr-Schild BH with BSSN, excision, radiative
boundary conditions and excision. An immediate CVS update is not for the faint of the heart -
I will do more testing in the next days.