this relatively minor change selectively changes the way SYNCS are done for Setter thorns:
If the Calculation structure has only one loop then SYNCS are done
in schedule.ccl, otherwise they are done in the source code as before,
and a warning message is printed.

This is supposed to be the minimal change necessary to adapt to the
Multipatch infrastructure. As long as Multipatch is not used, the user
does not need to care.