Added Errors module
Added query about excessive sorting

Renamed "derivativesUsed" to "oldDerivativesUsed" as we now have
two types of derivative.

Added error checking to make sure that only recognized functions are
used in a calculation and that shorthand names do not conflict with
gridfunction names. Ensure that shorthands are assigned before they
are used.

Fixed typo; a semicolon was missing in UncommentSourceSync. This code
could not possibly have worked for C thorns with more than one loop in
a calculation.

Ensure that the grid loop range is correct for *both* types of
derivatives, not just the old ones.

Removed check on whether or not derivatives are used when deciding
whether to synchronize. It makes no sense; we want to know if
*subsequent* loops will use derivatives, and that is too hard to
tell. So we now synchronize any variables which are set (either in the
source file or in the schedule file).