Reorganized the file into two parts.  The top section consists of functions which understand the internal layout of the Groups structure.  The remaining functions only use the functional interface. This was necessary as I needed to modify the Groups structure (in a backward compatible way) to include tensor type tags.  So now instead of
{"groupname", {var1, var2, ...}}

we can now have

{"groupname", {var1, var2,...}, Tags -> {TensorType -> "DD", TensorWeight -> -1/2}}

for example. The rest of Kranc should use ONLY the functions at the
top of this file to manipulate Groups structures. KrancThorn and
CalculationFunction will satisfy this requirement, but I am no longer
maintaining KrancThorns.m, so it won't be able to cope with the
modified group format.