Added Eps to sym context
Added functions for symmetrizing and
antisymmetrizing expressions in pairs of indices.

Added function to query whether an index is a lower or an upper index.

Invented the concept of a 'tensor attribute'.  This is a piece of
information about a particular tensor.  Set and get with
SetTensorAttribute[kernel,attr,value] and
GetTensorAttribute[kernel,attr]. Intention is to have things like
Symmetries, TensorWeight, TensorSpecial etc.  This allows these things
to be stored along with the tensor, and they can then be queried by
the parts of Kranc that understand tensors.  Symmetries are stored as
permutation weight lists as in MathTensor, though we still only have a
very rudimentary index symmetry system.

Changed MakeExplicit on a list of rules to allow duplicate entries.
This has the effect of allowing a shorthand or a grid function to be
set more than once in a calculation, which some people like to do
(i.e.  x = f[], x=g[x]).  Not sure why the duplicates were removed in
the first place, but was probably to get rid of extra equations caused
by the symmetry system.  We are now more clever about recognizing
tensors when this is called, so I hope it is no longer an issue.

Define symmetry tensor attribute when "AssertSymmetric*" are called.

Allow Eps to contain tensorial indices during validation check.

New function to determine the reflection symmetries of a tensor.