Added feature so that the user can loop over more points in a calculation than the stencil_width will allow.  This is useful if you know that your difference operators are actually safe to use there.
No longer "clean" the calculation five times in a row.  In fact, no
longer clean it at all.  What was this for? It looks like it was for
removing unused shorthands from a calculation.  But only those
shorthands which are used are declared anyway, so why is this
necessary? The logic used in cleancalc doesn't find the shorthands
which are used inside derivative operators, and this was stopping my
code from working.

Only precompute those derivatives that do not contain shorthands.  If
a derivative contains a shorthand, the shorthands needs to be computed
before the derivative, so precomputation is not really possible.