Added parameter for every calculation '_calc_every' and '_calc_offset' so that the calculation only happens if cctk_iteration % calc_every == calc_offset.
Fixed CollectList in KrancThorn.m (it was disabled).

Enabled MMA Simplify for expressions even when a CollectList has not
be specified.

Changed SYNC logic in KrancThorn.m: Calculations are now only SYNCed
if they are restricted to the interior of the grid (using
Where). Previously, all calculations were synced.  This is a
performance issue.  Further, calculations in MoL_CalcRHS and
MoL_RHSBoundaries are *never* SYNCed now.  The SYNC of the evolved
variables that always happens in cactus_boundaries.c should be enough.

WARNING: you probably want to test that this does not break your code.