Introduce a new grid function iterator InteriorNoSync.  This has the same effect as Interior, but does not synchronise the variables.  This can be used e.g. to calculate the RHS terms.
Introduce a way to access grid function that may or may not have storage.  If a local variable name matches the pattern "eT*L", then it is read from the corresponding grid function only if the parameter useMatter is true; otherwise it is set to zero.  This is obviously intended for the TmunuBase variables.  This is implemented via a new Kranc function maybeAssignVariableInLoop.