CalculationFunction.m: Replace Mathematica's line-breaking with our own
Mathematica's implementation of line-breaking in CForm sometimes
splits lines in the middle of tokens and adds continuation backslashes
to the end of the split line.  In combination with the indentation
that we apply in Kranc, this can cause the tokens to be broken by
whitespace.  This commit inhibits Mathematica's line breaking and
replaces it with an algorithm which splits lines only on whitespace.