CalculationFunction.m: Remove ability to sync groups in the source file
This feature was never really used, and was only introduced before we
really understood what we were doing.  The idea was that you could
have multiple loops in a single calculation and you might have to sync
between them.  For example, in the first loop you may compute some
quantity, and in the second you might compute its derivative.  This
would have to be in the second loop because otherwise it would lead to
incorrect results.  However, this scheme only allows synchronisation,
it does not take account of symmetry, interpatch or physical boundary
conditions, and possibly doesn't work correctly with mesh refinement.
Anything which can be done with this scheme can be done simply by
using a second calculation, so in the interests of maintainability and
code conciseness, the feature has been removed.