Remove support for original finite differencing implementation
The original implementation of finite differencing in Kranc used
hard-coded finite difference operator macros in GenericFD.  It has not
been tested in years and probably doesn't even work any more.  The new
implementation allows the user to create arbitrary operators and can
do everything that the old one could.  As we are planning to replace
the current implementation with something more user-friendly, to
simplify the code we are removing the old method at this point.