Remove support for sub-block grid functions
The idea of sub-block grid functions was to allow a calculation to
operate differently on different sub-blocks of the domain.  This was
never used, and is now achieved by calling the calculation with
arguments which specify the region of the domain to loop over.  Hence,
the old sub-block code is being removed.