Add automatic scheduling of calculations in MoL_PseudoEvolution
Before this commit, Kranc required the user to use the Schedule key in
a calculation to give a Cactus schedule specification.  Scheduling
analysis functions which compute only on the grid interior (e.g. those
that take derivatives) was problematic because boundary conditions
were not applied (e.g. symmetries and physical), leading to either
poison or undefined values on the grid.  Calculations scheduled in
ANALYSIS could not have prolongation in time with mesh refinement.

This commit introduces a new way to schedule functions which is
completely automatic.  If the Schedule key is omitted, the calculation
is scheduled in MoL_PseudoEvolution which means it will be called in
all necessary places.  Each such calculation will have its own
schedule group, in which boundary conditions are also selected and
applied (if necessary).  Boundary conditions are also applied
in POSTRESTRICT and POSTRESTRICTINITIAL.