Support no-inline functions for derivatives in vectorised code
Create two versions of the derivative operators, once as macro, and
once as no-inline function. When a function is used, then the grid
spacing 1/dx must be passed in explicitly. This was done differently
before and is now slightly faster and shorter. The no-inline function
reduces code size, and should probably also be available for the
non-vectorised code, but this is not implemented.