Correct vectorisation errors recently introduced
(1) vec_load cannot be used on constants, it can only be used to
access array elements. Replacing UseVectors by False is just a band
aid. vec_load was attached to a routine defining and setting a
variable; really, it should be attached to a routine accessing array
elements, but Kranc doesn't use such an abstraction yet.(2) ToReal is
introduced, but must be removed again for integer expressions such as
the conditions controlling if statements. I added band-aid code to
Conditional[] to remove it. The vectorization routines already handle
IfThen, Pow etc., but Conditional isn't visible to them.