Jacobian.m: Insert Jacobian multiplications after any shorthands that are used in the derivatives
In order to multiply derivatives by the Jacobians, it is necessary to
evaluate the derivative operators.  If shorthands are required for
this, they must be computed before the Jacobians are multiplied.  With
this commit, Kranc will compute the Jacobianised derivatives as early
as possible in the calculation after all the shorthands used in them
have been assigned, rather than computing them at the start of the
calculation.