wave_sine.par: Increase ghost and boundary size to 2
This is necessary because Kranc checks that you have enough ghost and boundary points for all the finite difference operators you use, and the boundary condition calculation uses PDonesided2nd, which requires a ghost size of 2.  Given how this is used, it almost certainly can be used with a smaller ghost size, but the check in Kranc does not know this.