Add and use GetCalculationScheduleName
A calculation has a name, stored in its Name entry.  This is what is used when generating the code files, and is what users can use to do relative scheduling (before, after, etc).  Sometimes (e.g. for CaKernel), the routine scheduled with the Cactus scheduler in  schedule.ccl needs to have a different name.  This is what is returned by GetCalculationScheduleName.  We now always schedule such functions "<schedulename> as <name>" where <name> is the real name of the calculation.