Calculate mixed derivatives as repeated first derivatives
This is controlled by a new calculation option SeparatedDerivatives2.  It reduces the number of floating point operations and converts a large 2D stencil into a 1D stencil.  This is possible when the first derivatives are stored in grid functions, and then they are computed in ghost zones.  This is currently only true for CaKernel, not for the standard CPU code.