GenericFD: Correct implementation of Sign
Mathematica defines Sign(x) = Which[x>0, 1, x<0, -1, x==0, 0] whereas
the current implementation in terms of copysign gives Sign(0) != 0.
This causes problems when Sign is used to choose the finite
differencing operator in boundary conditions.  This commit restores
the old definition in terms of the sgn function which we define now as
static inline in GenericFD.h.