Use CCTK_ATTRIBUTE_UNUSED to mark declared variables as unused
In automatically-generated code, it is often convenient to declare variables unconditionally, and let the compiler strip the unused definitions out.  Eventually it would be good to tidy up Kranc so that it only declares variables which are actually needed, but for the moment it is more important to reduce the number of warnings generated during compilation, to encourage users to read the warnings that might actually indicate a problem.