Improve handling of LoopControl
LoopControl support is enabled for a thorn and all calculations by default.  It can be disabled for individual calculations by adding UseLoopControl -> False to the calculation.  It can be disabled for the whole thorn by passing UseLoopControl -> False to CreateThorn.  Use of LoopControl is required if UseVectors is True because the vectorisation code uses some features from LoopControl.