Introduce targets for CodeGen
CodeGen targets currently include C (with or without vectorisation) and CaKernel.  The target is an Object which can have arbitrary keys.  Currently the target is propagated through the code using a global variable $CodeGenTarget, but eventually this will be replaced with passing to functions as an argument.  The current vectorise argument will then be redundant. Additionally, introduce a PostProcessExpression function provided by the target.  This is applied after "processing" and before code generation.