Thorn.m: Catch errors when merging files
If a broken symbolic link, e.g. as created by Emacs as a lock file, is
found in the merge directory, Import generates a message which doesn't
indicate the problematic filename and continues, and Kranc chokes on the
result.  In this commit, check for messages generated during Import, and
if there were any, abort with a fatal error with the filename in the
error message.

See #115.