TensorTools.m: Work around bug in dummy index relabelling
There is some as-yet-unidentified problem with the relabelling of dummy
indices which appears in complicated expressions involving 2nd covariant
derivatives of tensors.  Instead of relying on the relabelling code,
determine a single pool of unused indices in MakeExplicit, and use each
one only once when introducing new dummy indices. This guarantees that
there will be no conflicts, but is more likely to run out of indices
when the pool is empty.  In that case, a fatal error will be reported,
and we will need to either fix the relabelling bug or introduce more
indices.